requirements.txt
streamlit
pandas
openpyxl
requests
unidecode
feedparser
Pillow
